renaming trackCategoryLookup folder to track-category
